+++ EMS/admin-ui-bundle/assets/js/core/plugins/sortableList.js
@@ -1,0 +1,7 @@
+import $ from 'jquery'
+
+export default class SortableList {
+  load (target) {
+    $(target).find('ul.sortable').sortable()
+  }
+}
